requirements.txt
openai
python-dotenv
rich
pysocks
